feat: 로그인 시 내 게시글 조회 및 회원 탈퇴 soft delete 적용

+++ src/main/java/com/practice/logincrud/board/BoardDto.java
@@ -6,8 +6,8 @@ import java.time.LocalDateTime;
 
 @Data
 public class BoardDto {
-    private Integer id;
-    private Integer memberId;
+    private Long id;
+    private Long memberId;
     private String title;
     private String content;
     private String nickName;
